https://contyads.com/?rlnk=jhodimalao
How do I make $ 300-800 a month?
I'm sure each of you asked yourself this question ...

We ContyAds have developed and calculated a strategy, a plan that works, a plan that has already brought success to many people. Our numbers are real numbers.
You can easily earn money with ContyAds, but you should work for it. Without doing nothing, you cannot earn anything in life!
Here's a little calculation:
You invite 5 people to buy advertising on ContyAds, in return you get 10[percent] of the advertising costs ALWAYS for EVERY advertising when the advertiser buys advertising. And he will buy a lot and very often advertising that is guaranteed. An advertiser orders for around $ 300-600 a month if you get 10[percent] that's $ 30-60 per person. But when you have 5 or 10 advertisers ???
You get the 10[percent] commission on 3 levels, if your partners do the same then you also get 10[percent] commission.
So $ 300-800 a month is very easy to make!
Everyone should just make an effort ...

We have now also reduced advertising costs, for example for the 728x90 banner you can now get three times more impressions for the same money. Or for all other advertising options you can already buy advertising for 200 visits and not for 1000 as it was before.

IMPORTANT!
Once we reach 60,000-80,000 partners, we will reward our members. All members receive a share of the ContyAds profit and receive a certain percentage of the profit. That can be 30-70 $ a month that we will pay out to each partner. Our partner company did the same 3 years ago and has been very successful with it to this day. This company already has over 4,000,000 members. They share their profit with the members and everyone earns and is satisfied.

In the next few months there will be a lot more innovations at ContyAds! We will build our success together with you, and you will benefit greatly from it and make profits.